+++ src/ScanLib.h
@@ -129,12 +129,12 @@ typedef int (*SL_SCAN_IMAGE)(unsigned verbose, unsigned dpi, double left,
 EXPORT int
 slDecodePlate(unsigned verbose, unsigned dpi, int brightness, int contrast,
 		unsigned plateNum, double left, double top, double right,
-		double bottom, unsigned scanGap, unsigned squareDev,
+		double bottom, double scanGap, unsigned squareDev,
 		unsigned edgeThresh, unsigned corrections, double cellDistance);
 
 typedef int (*SL_DECODE_PLATE)(unsigned verbose, unsigned dpi, int brightness,
 		int contrast, unsigned plateNum, double left, double top, double right,
-		double bottom, unsigned scanGap, unsigned squareDev,
+		double bottom, double scanGap, unsigned squareDev,
 		unsigned edgeThresh, unsigned corrections, double cellDistance);
 
 /**
@@ -176,11 +176,11 @@ typedef int (*SL_DECODE_PLATE)(unsigned verbose, unsigned dpi, int brightness,
  *  not be determined.
  */
 EXPORT int slDecodeImage(unsigned verbose, unsigned plateNum, char * filename,
-		unsigned scanGap, unsigned squareDev, unsigned edgeThresh,
+		double scanGap, unsigned squareDev, unsigned edgeThresh,
 		unsigned corrections, double cellDistance);
 
 typedef int (*SL_DECODE_IMAGE)(unsigned verbose, unsigned plateNum,
-		char * filename, unsigned scanGap, unsigned squareDev,
+		char * filename, double scanGap, unsigned squareDev,
 		unsigned edgeThresh, unsigned corrections, double cellDistance);
 
 #ifdef __cplusplus
